+++ modular/leaderboard/x/leaderboard/types/packet_candidate.go
@@ -1,18 +1,13 @@
 package types
 
-import (
-	"errors"
-)
-
 // ValidateBasic is used for validating the packet
 func (p CandidatePacketData) ValidateBasic() error {
 
-	// TODO: Validate the packet data
-
-  // return error if player address is empty
-  if p.PlayerInfo.Index == "" {
-      return errors.New("Player address cannot be empty")
-  }
+	// return error if player info is incorrect
+	playerInfoErr := p.PlayerInfo.ValidateBasic()
+	if playerInfoErr != nil {
+		return playerInfoErr
+	}
 
 	return nil
 }
