+++ modular/leaderboard/x/leaderboard/types/player_info.go
@@ -1,0 +1,13 @@
+package types
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+func (info PlayerInfo) ValidateBasic() error {
+	_, err := sdk.AccAddressFromBech32(info.Index)
+	if err != nil {
+		return err
+	}
+	return nil
+}
